fixing platform dependency of node ids (some longs are longer than others), refs #21
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@24022 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/common/IDSupplier.cpp
@@ -39,7 +39,7 @@
 // ===========================================================================
 // method definitions
 // ===========================================================================
-IDSupplier::IDSupplier(const std::string& prefix, long begin)
+IDSupplier::IDSupplier(const std::string& prefix, long long int begin)
     : myCurrent(begin), myPrefix(prefix) {}
 
 
@@ -67,7 +67,7 @@ void
 IDSupplier::avoid(const std::string& id) {
     // does it start with prefix?
     if (id.find(myPrefix) == 0) {
-        long number;
+        long long int number;
         std::istringstream buf(id.substr(myPrefix.size()));
         buf >> number;
         // does it continue with a number?
